UPSTREAM: <carry>: openshift: move all test utility functions
Moved all test utility & setup functions into
machineapi_controller_test.go.

In the vast majority of the unit tests you need a controller so it
makes sense to consolidate all the utilities in that file.
