+++ cluster-autoscaler/cloudprovider/openshiftmachineapi/machineapi_nodegroup_test.go
@@ -23,11 +23,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/openshift/cluster-api/pkg/apis/machine/v1beta1"
 	apiv1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/autoscaler/cluster-autoscaler/cloudprovider"
 	"k8s.io/utils/pointer"
 )
@@ -35,206 +33,6 @@ import (
 const (
 	machineAnnotationKey = "machine.openshift.io/machine"
 )
-
-type clusterTestConfig struct {
-	spec              *clusterTestSpec
-	machineDeployment *v1beta1.MachineDeployment
-	machineSet        *v1beta1.MachineSet
-	machines          []*v1beta1.Machine
-	nodes             []*apiv1.Node
-}
-
-type clusterTestSpec struct {
-	annotations             map[string]string
-	id                      int
-	machineDeploymentPrefix string
-	machineSetPrefix        string
-	namespace               string
-	nodeCount               int
-	replicaCount            int32
-	rootIsMachineDeployment bool
-}
-
-func (config clusterTestConfig) newNodeGroup(t *testing.T, c *machineController) (*nodegroup, error) {
-	if config.machineDeployment != nil {
-		return newNodegroupFromMachineDeployment(c, config.machineDeployment)
-	}
-	return newNodegroupFromMachineSet(c, config.machineSet)
-}
-
-func (config clusterTestConfig) newMachineController(t *testing.T) (*machineController, testControllerShutdownFunc) {
-	nodeObjects := make([]runtime.Object, len(config.nodes))
-	machineObjects := make([]runtime.Object, len(config.machines))
-
-	for i := range config.nodes {
-		nodeObjects[i] = config.nodes[i]
-	}
-
-	for i := range config.machines {
-		machineObjects[i] = config.machines[i]
-	}
-
-	machineObjects = append(machineObjects, config.machineSet)
-	if config.machineDeployment != nil {
-		machineObjects = append(machineObjects, config.machineDeployment)
-	}
-
-	return mustCreateTestController(t, testControllerConfig{
-		nodeObjects:    nodeObjects,
-		machineObjects: machineObjects,
-	})
-}
-
-func newMachineSetTestObjs(namespace string, id, nodeCount int, replicaCount int32, annotations map[string]string) *clusterTestConfig {
-	spec := &clusterTestSpec{
-		id:                      id,
-		annotations:             annotations,
-		machineSetPrefix:        fmt.Sprintf("machineset-%s-", namespace),
-		namespace:               namespace,
-		nodeCount:               nodeCount,
-		replicaCount:            replicaCount,
-		rootIsMachineDeployment: false,
-	}
-
-	return makeClusterObjs(spec)
-}
-
-func newMachineDeploymentTestObjs(namespace string, id, nodeCount int, replicaCount int32, annotations map[string]string) *clusterTestConfig {
-	spec := &clusterTestSpec{
-		id:                      id,
-		annotations:             annotations,
-		machineDeploymentPrefix: fmt.Sprintf("machinedeployment-%s-", namespace),
-		machineSetPrefix:        fmt.Sprintf("machineset-%s-", namespace),
-		namespace:               strings.ToLower(namespace),
-		nodeCount:               nodeCount,
-		replicaCount:            replicaCount,
-		rootIsMachineDeployment: true,
-	}
-
-	return makeClusterObjs(spec)
-}
-
-func makeClusterObjs(spec *clusterTestSpec) *clusterTestConfig {
-	objs := clusterTestConfig{
-		spec:     spec,
-		nodes:    make([]*apiv1.Node, spec.nodeCount),
-		machines: make([]*v1beta1.Machine, spec.nodeCount),
-	}
-
-	objs.machineSet = &v1beta1.MachineSet{
-		TypeMeta: v1.TypeMeta{
-			Kind: "MachineSet",
-		},
-		ObjectMeta: v1.ObjectMeta{
-			Name:      fmt.Sprintf("%s%d", spec.machineSetPrefix, spec.id),
-			Namespace: spec.namespace,
-			UID:       types.UID(fmt.Sprintf("%s%d", spec.machineSetPrefix, spec.id)),
-		},
-	}
-
-	if !spec.rootIsMachineDeployment {
-		objs.machineSet.ObjectMeta.Annotations = spec.annotations
-		objs.machineSet.Spec.Replicas = int32ptr(spec.replicaCount)
-	} else {
-		objs.machineDeployment = &v1beta1.MachineDeployment{
-			TypeMeta: v1.TypeMeta{
-				Kind: "MachineDeployment",
-			},
-			ObjectMeta: v1.ObjectMeta{
-				Name:        fmt.Sprintf("%s%d", spec.machineDeploymentPrefix, spec.id),
-				Namespace:   spec.namespace,
-				UID:         types.UID(fmt.Sprintf("%s%d", spec.machineDeploymentPrefix, spec.id)),
-				Annotations: spec.annotations,
-			},
-			Spec: v1beta1.MachineDeploymentSpec{
-				Replicas: int32ptr(spec.replicaCount),
-			},
-		}
-
-		objs.machineSet.OwnerReferences = make([]v1.OwnerReference, 1)
-		objs.machineSet.OwnerReferences[0] = v1.OwnerReference{
-			Name: objs.machineDeployment.Name,
-			Kind: objs.machineDeployment.Kind,
-			UID:  objs.machineDeployment.UID,
-		}
-	}
-
-	machineOwner := v1.OwnerReference{
-		Name: objs.machineSet.Name,
-		Kind: objs.machineSet.Kind,
-		UID:  objs.machineSet.UID,
-	}
-
-	for i := 0; i < spec.nodeCount; i++ {
-		objs.nodes[i], objs.machines[i] = makeLinkedNodeAndMachine(i, spec.namespace, machineOwner)
-	}
-
-	return &objs
-}
-
-func int32ptr(v int32) *int32 {
-	return &v
-}
-
-func makeMachineSet(i int, replicaCount int, annotations map[string]string) *v1beta1.MachineSet {
-	return &v1beta1.MachineSet{
-		TypeMeta: v1.TypeMeta{
-			Kind: "MachineSet",
-		},
-		ObjectMeta: v1.ObjectMeta{
-			Name:        fmt.Sprintf("machineset-%d", i),
-			Namespace:   "test-namespace",
-			UID:         types.UID(fmt.Sprintf("machineset-%d", i)),
-			Annotations: annotations,
-		},
-		Spec: v1beta1.MachineSetSpec{
-			Replicas: int32ptr(int32(replicaCount)),
-		},
-	}
-}
-
-// makeLinkedNodeAndMachine creates a node and machine. The machine
-// has its NodeRef set to the new node and the new machine's owner
-// reference is set to owner.
-func makeLinkedNodeAndMachine(i int, namespace string, owner v1.OwnerReference) (*apiv1.Node, *v1beta1.Machine) {
-	node := &apiv1.Node{
-		TypeMeta: v1.TypeMeta{
-			Kind: "Node",
-		},
-		ObjectMeta: v1.ObjectMeta{
-			Name: fmt.Sprintf("node-%d", i),
-			Annotations: map[string]string{
-				machineAnnotationKey: fmt.Sprintf("%s/machine-%d", namespace, i),
-			},
-		},
-		Spec: apiv1.NodeSpec{
-			ProviderID: fmt.Sprintf("nodeid-%d", i),
-		},
-	}
-
-	machine := &v1beta1.Machine{
-		TypeMeta: v1.TypeMeta{
-			Kind: "Machine",
-		},
-		ObjectMeta: v1.ObjectMeta{
-			Name:      fmt.Sprintf("machine-%d", i),
-			Namespace: namespace,
-			OwnerReferences: []v1.OwnerReference{{
-				Name: owner.Name,
-				Kind: owner.Kind,
-				UID:  owner.UID,
-			}},
-		},
-		Status: v1beta1.MachineStatus{
-			NodeRef: &apiv1.ObjectReference{
-				Kind: node.Kind,
-				Name: node.Name,
-			},
-		},
-	}
-
-	return node, machine
-}
 
 func TestNodeGroupNewNodeGroup(t *testing.T) {
 	type testCase struct {
